% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hubStyle-package.R
\docType{package}
\name{hubStyle-package}
\alias{hubStyle}
\alias{hubStyle-package}
\title{hubStyle: pkgdown template and utilities for hubverse docs}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This is a private template for use by hubverse packages. Please don't use it for your own non-hubverse package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hubverse-org/hubStyle}
  \item Report bugs at \url{https://github.com/hubverse-org/hubStyle/issues}
}

}
\author{
\strong{Maintainer}: Anna Krystalli \email{annakrystalli@googlemail.com} (\href{https://orcid.org/0000-0002-2378-4915}{ORCID})

}
\keyword{internal}
